refactor: Make delete command follow the other commands pattern

+++ internal/k8sbox/handlers/deleteHandler.go
@@ -8,10 +8,16 @@ import (
 	"strings"
 
 	model "github.com/twelvee/k8sbox/internal/k8sbox/models"
+	"github.com/twelvee/k8sbox/pkg/k8sbox/structs"
+	"k8s.io/utils/strings/slices"
 )
 
 // HandleDeleteCommand is the k8sbox delete command handler
-func HandleDeleteCommand(context context.Context, tomlFile string, environmentID string) {
+func HandleDeleteCommand(context context.Context, modelName string, tomlFile string, environmentID string) {
+	if !slices.Contains(structs.GetEnvironmentAliases(), modelName) {
+		fmt.Printf("Invalid argument. Available types: %s\r\n", strings.Join(structs.GetEnvironmentAliases(), ", "))
+		os.Exit(1)
+	}
 	if len(strings.TrimSpace(tomlFile)) != 0 {
 		err := model.DeleteEnvironmentByTomlFile(tomlFile)
 		if err != nil {
